uf-4474: Fixed lift issues and bumped dept44.

+++ src/main/java/se/sundsvall/myrepresentative/service/authorities/AuthoritiesResponseMapper.java
@@ -32,7 +32,7 @@ public class AuthoritiesResponseMapper {
         return MetaData.builder()
                 .withPage(response.getPage().getNumber())
                 .withTotalPages(response.getPage().getTotalPages().intValue())
-                .withTotalRecords(response.getPage().getTotalElements().intValue())
+                .withTotalRecords(response.getPage().getTotalElements())
                 .withLimit(response.getPage().getSize())
                 .build();
     }
